moved chunked graph code into graphene datasource and combined some classes
