+++ src/neuroglancer/datasource/graphene/base.ts
@@ -16,6 +16,8 @@
 
 import {mat4} from 'neuroglancer/util/geom';
 import {ShardingParameters} from 'neuroglancer/datasource/precomputed/base';
+import {ChunkLayoutOptions, makeSliceViewChunkSpecification, SliceViewChunkSource, SliceViewChunkSpecification, SliceViewChunkSpecificationBaseOptions, SliceViewChunkSpecificationOptions} from 'neuroglancer/sliceview/base';
+import {DataType} from 'neuroglancer/sliceview/base';
 
 export const PYCG_APP_VERSION = 1;
 
@@ -74,4 +76,53 @@ export function getGrapheneFragmentKey(fragmentId: string) {
   } else {
     return {key:fragmentId, fragmentId: fragmentId};
   }
+}
+
+export const CHUNKED_GRAPH_LAYER_RPC_ID = 'ChunkedGraphLayer';
+export const CHUNKED_GRAPH_RENDER_LAYER_UPDATE_SOURCES_RPC_ID = 'ChunkedGraphLayer:updateSources'
+export const RENDER_RATIO_LIMIT = 5.0;
+
+export interface ChunkedGraphChunkSpecificationBaseOptions extends
+    SliceViewChunkSpecificationBaseOptions {
+  /**
+   * Specifies offset for use by backend.ts:GenericVolumeChunkSource.computeChunkBounds in
+   * calculating chunk voxel coordinates.  The calculated chunk coordinates will be equal to the
+   * voxel position (in chunkLayout coordinates) plus this value.
+   *
+   * Defaults to kZeroVec if not specified.
+   */
+  baseVoxelOffset?: Float32Array;
+  dataType: DataType;
+}
+
+export interface ChunkedGraphChunkSpecificationOptions extends
+    ChunkedGraphChunkSpecificationBaseOptions, SliceViewChunkSpecificationOptions<Uint32Array> {}
+
+/**
+ * Specifies parameters for ChunkedGraphChunkSpecification.getDefaults.
+ */
+export interface ChunkedGraphChunkSpecificationGetDefaultsOptions extends
+    ChunkedGraphChunkSpecificationBaseOptions, ChunkLayoutOptions {}
+
+/**
+ * Specifies a chunk layout and voxel size.
+ */
+export interface ChunkedGraphChunkSpecification extends SliceViewChunkSpecification<Uint32Array> {
+  baseVoxelOffset: Float32Array;
+  dataType: DataType;
+}
+
+export function makeChunkedGraphChunkSpecification(options: ChunkedGraphChunkSpecificationOptions): ChunkedGraphChunkSpecification {
+  const {rank, dataType} = options;
+  const {baseVoxelOffset = new Float32Array(rank)} = options;
+
+  return {
+    ...makeSliceViewChunkSpecification(options),
+    baseVoxelOffset,
+    dataType,
+  }
+}
+
+export interface ChunkedGraphChunkSource extends SliceViewChunkSource {
+  spec: ChunkedGraphChunkSpecification;
 }
